+++ src/pages/droner/EditDroner.tsx
@@ -42,6 +42,7 @@ import {
 } from "../../entities/DronerDroneEntities";
 import ModalDrone from "../../components/modal/ModalDronerDrone";
 import {
+  DRONER_DRONE_MAPPING,
   DRONER_DRONE_STATUS,
   DRONER_STATUS,
   STATUS_COLOR,
@@ -1002,7 +1003,7 @@ function EditDroner() {
                   <div className="col-lg-4">
                     <span style={{ color: STATUS_COLOR[item.status] }}>
                       <Badge color={STATUS_COLOR[item.status]} />
-                      {DRONER_DRONE_STATUS[item.status]}
+                      {DRONER_DRONE_MAPPING[item.status]}
                       <br />
                     </span>
                   </div>